[refactor] : 함수 import 및 실행으로 구조 변경 window.onload 제거
